Use size_t whenever counting the number of objects

+++ minisat/core/Main.cc
@@ -147,7 +147,7 @@ int main(int argc, char** argv)
         if (res != NULL){
             if (ret == l_True){
                 fprintf(res, "s SATISFIABLE\n");
-                for (int i = 0; i < S.nVars(); i++)
+                for (size_t i = 0; i < S.nVars(); i++)
                     if (S.model[i] != l_Undef)
                         fprintf(res, "%s%s%d", (i==0)?"":" ", (S.model[i]==l_True)?"":"-", i+1);
                 fprintf(res, " 0\n");
