+++ minisat/mtl/Alloc.h
@@ -33,11 +33,11 @@ template<class T>
 class RegionAllocator
 {
     T*        memory;
-    uint32_t  sz;
-    uint32_t  cap;
-    uint32_t  wasted_;
+    size_t    sz;
+    size_t    cap;
+    size_t    wasted_;
 
-    void capacity(uint32_t min_cap);
+    void capacity(size_t min_cap);
 
  public:
     // TODO: make this a class for better type-checking?
@@ -45,7 +45,7 @@ class RegionAllocator
     enum { Ref_Undef = UINT32_MAX };
     enum { Unit_Size = sizeof(T) };
 
-    explicit RegionAllocator(uint32_t start_cap = 1024*1024) : memory(NULL), sz(0), cap(0), wasted_(0){ capacity(start_cap); }
+    explicit RegionAllocator(size_t start_cap = 1024*1024) : memory(NULL), sz(0), cap(0), wasted_(0){ capacity(start_cap); }
     ~RegionAllocator()
     {
         if (memory != NULL)
@@ -53,11 +53,11 @@ class RegionAllocator
     }
 
 
-    uint32_t size      () const      { return sz; }
-    uint32_t wasted    () const      { return wasted_; }
+    size_t   size      () const      { return sz; }
+    size_t   wasted    () const      { return wasted_; }
 
-    Ref      alloc     (int size); 
-    void     free      (int size)    { wasted_ += size; }
+    Ref      alloc     (size_t size);
+    void     free      (size_t size)    { wasted_ += size; }
 
     // Deref, Load Effective Address (LEA), Inverse of LEA (AEL):
     T&       operator[](Ref r)       { assert(r < sz); return memory[r]; }
@@ -83,17 +83,17 @@ class RegionAllocator
 };
 
 template<class T>
-void RegionAllocator<T>::capacity(uint32_t min_cap)
+void RegionAllocator<T>::capacity(size_t min_cap)
 {
     if (cap >= min_cap) return;
 
-    uint32_t prev_cap = cap;
+    size_t prev_cap = cap;
     while (cap < min_cap){
         // NOTE: Multiply by a factor (13/8) without causing overflow, then add 2 and make the
         // result even by clearing the least significant bit. The resulting sequence of capacities
         // is carefully chosen to hit a maximum capacity that is close to the '2^32-1' limit when
         // using 'uint32_t' as indices so that as much as possible of this space can be used.
-        uint32_t delta = ((cap >> 1) + (cap >> 3) + 2) & ~1;
+        size_t delta = ((cap >> 1) + (cap >> 3) + 2) & ~((size_t)1);
         cap += delta;
 
         if (cap <= prev_cap)
@@ -108,13 +108,13 @@ void RegionAllocator<T>::capacity(uint32_t min_cap)
 
 template<class T>
 typename RegionAllocator<T>::Ref
-RegionAllocator<T>::alloc(int size)
+RegionAllocator<T>::alloc(size_t size)
 { 
     // printf("ALLOC called (this = %p, size = %d)\n", this, size); fflush(stdout);
     assert(size > 0);
     capacity(sz + size);
 
-    uint32_t prev_sz = sz;
+    size_t prev_sz = sz;
     sz += size;
     
     // Handle overflow:
